got dash working now need to make it look nice and simple
